several optimizations following PR discussion
