+++ src/modules/mock.js
@@ -1,10 +1,6 @@
 var $ = require('../core/init');
 
-var m = {
-	deps: function(mock) {
-		$.initialize(window = mock || window);
-		return window;
-	}
+exports.deps = function(mock) {
+	$.initialize(window = mock || window);
+	return window;
 }
-
-module.exports = m;
